pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phoenix-observability"
version = "0.1.0"
description = "Unified LLM Observability Package for Phoenix + OpenTelemetry"
authors = [{  }]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["observability", "phoenix", "opentelemetry", "llm", "monitoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp>=1.25",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.5",
]

[project.optional-dependencies]
phoenix = [
    "arize-phoenix>=2.5.0",
]

[project.urls]
Homepage = "https://github.com/your-org/phoenix-observability"
Documentation = "https://github.com/your-org/phoenix-observability#readme"
Repository = "https://github.com/your-org/phoenix-observability"

[tool.setuptools]
packages = ["phoenix_observability"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]

